[PRM-SanitiseFileName] - Added feature switch for sanitising file name + moved creating SDES notifications to service + test changes
