+++ conf/application.conf
@@ -139,6 +139,7 @@ feature {
     call-api-1811-etmp = false
     time-machine-now = ""
     use-internal-auth = false
+    sanitise-file-name = true
   }
 }
